requirements.txt
# SENTINEL AI-SOC — Python dependencies
# Generated from actual imports in the codebase

# Web framework
fastapi==0.115.6
uvicorn==0.34.0
starlette==0.41.3
websockets==15.0.1
python-multipart==0.0.20

# Pydantic (data validation / settings)
pydantic==2.10.4
pydantic-settings==2.7.1
pydantic_core==2.27.2
annotated-types==0.7.0

# Database
SQLAlchemy==2.0.36

# AI / ML
scikit-learn==1.8.0
numpy==2.2.6
joblib==1.5.2
scipy==1.17.1
threadpoolctl==3.6.0

# NLP — DistilBERT phishing classifier
transformers==5.3.0
torch==2.10.0
tokenizers==0.22.2
huggingface_hub==1.7.2
safetensors==0.7.0
filelock==3.20.0
tqdm==4.67.1
regex==2026.2.28
fsspec==2026.2.0

# Google Gemini AI analyst (new SDK — replaces deprecated google-generativeai)
google-genai==1.59.0

# Twilio SMS alerts
twilio==9.10.3

# HTTP client (AbuseIPDB + external API calls)
httpx==0.28.1
httpcore==1.0.9
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.0
idna==3.10

# YAML (playbooks)
PyYAML==6.0.3

# Faker (event simulation)
Faker==40.11.0

# Async
anyio==4.10.0
sniffio==1.3.1
h11==0.16.0

# Env vars
python-dotenv==1.0.1

# Utilities
python-dateutil==2.9.0.post0
six==1.17.0
packaging==25.0
typing_extensions==4.15.0
